pyproject.toml
[project]
name = "gnma-data-manager"
version = "0.1.0"
description = "Tools for managing the GNMA public disclosure datasets."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyarrow>=19.0.1",
    "python-decouple>=3.8",
    "ruff>=0.11.10",
    "polars>=1.17.0",
]
